pe-portfolio-monitor/backend/pyproject.toml
[project]
name = "pe-portfolio-monitor"
version = "0.1.0"
description = "PE Portfolio Monitoring Dashboard"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "fastapi-users[fastapi]>=13.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pyotp>=2.9.0",
    "python-multipart>=0.0.17",
    "aiocsv>=1.1.0",
    "httpx>=0.28.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
